// src/api/generated/model/calculationQuestion.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionDifficulty } from './questionDifficulty';
import type { CalculationDetails } from './calculationDetails';

/**
 * Model for calculation questions.
 */
export interface CalculationQuestion {
  question_id?: string;
  text: string;
  difficulty: QuestionDifficulty;
  knowledge_node_id: string;
  question_type?: 'calculation';
  details: CalculationDetails;
}