// src/api/generated/model/fillInTheBlankQuestion.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionDifficulty } from './questionDifficulty';
import type { FillInTheBlankDetails } from './fillInTheBlankDetails';

/**
 * Model for fill in the blank questions.
 */
export interface FillInTheBlankQuestion {
  question_id?: string;
  text: string;
  difficulty: QuestionDifficulty;
  knowledge_node_id: string;
  question_type?: 'fill_blank';
  details: FillInTheBlankDetails;
}