// src/api/generated/model/multipleChoiceQuestion.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { QuestionDifficulty } from './questionDifficulty';
import type { MultipleChoiceDetails } from './multipleChoiceDetails';

/**
 * Model for multiple choice questions.
 */
export interface MultipleChoiceQuestion {
  question_id?: string;
  text: string;
  difficulty: QuestionDifficulty;
  knowledge_node_id: string;
  question_type?: 'multiple_choice';
  details: MultipleChoiceDetails;
}